server/requirements.txt
pytest
black
pytest-django
asgiref
attrs
autopep8
certifi
charset-normalizer
coreapi
coreschema
Django
django-cors-headers
djangorestframework
drf-yasg
idna
inflection
itypes
Jinja2
jsonschema
MarkupSafe
packaging
pycodestyle
pyparsing
pyrsistent
pytz
requests
ruamel.yaml
sqlparse
swagger-spec-validator
tomli
typing_extensions
tzdata
uritemplate
urllib3
setuptools